msm: kgsl: Avoid gpu boot if it is already up
Consider a scenario when touch wakeup happened before first open.
In such case, touch wakeup thread will turn on the GPU. Currently,
when first open happens, KGSL blindly goes ahead to turn on the GPU.
This will result in inappropriate system state. To avoid this, ensure
that KGSL does not attempt to turn on GPU if it is already on.

Change-Id: I8b797e848d5c97e54608e68f05ff98980177848e